added buggy admin panel

+++ utils/db_api/models.py
@@ -5,7 +5,7 @@ from utils.db_api.database import db
 
 class User(db.Model):
     __tablename__ = 'user'
-    chat_id = Column(Integer, primary_key=True)
+    chat_id = Column(BIGINT, primary_key=True)
     full_name = Column(String)
 
     def __repr__(self):
@@ -48,7 +48,7 @@ class Link(db.Model):
 class DateTimeForLink(db.Model):
     __tablename__ = 'datetime_for_link'
     id = Column(Integer, primary_key=True, autoincrement=True)
-    link_id = Column(ForeignKey("link.id"))  # переименовать links_id на link_id
+    link_id = Column(ForeignKey("link.id"))
     date = Column(Date)
     time_start = Column(Time)
     time_end = Column(Time)
